Fix 12 real-world failure modes identified by GNC, embedded-C, and V&V review
GNC algorithm fixes:
- attitude.c/h: add ATT_CalibrateGyroBias() — gyro bias was never estimated
  during pad calibration; a ±1°/s offset integrated over 4.2 s boost = ~4°
  pitch error, which combined with accel bias drift could push worst-case
  apogee detection past 100 ms
- apogee_detector.c: C1 counter now decrements (not resets) on a noise spike;
  a single sample above threshold no longer erases all prior negative-velocity
  evidence, making C1 robust to isolated transients
- apogee_detector.h: mark apogee_fired volatile — written from ISR-triggered
  processing, read from main loop; without volatile the compiler may cache the
  value in a register and the main loop never sees the fired state
- flight_fsm.c/h: add COAST→BOOST chuff recovery transition; if
  thrust-level acceleration reappears in COAST the motor is still burning
  (false burnout from a pressure oscillation > BURNOUT_DEBOUNCE_MS);
  remove unused velocity parameter from FSM_Update (MISRA Rule 2.7)

Embedded-C safety fixes:
- main.c: replace critical_enter/exit stubs with real MRS/CPSID/MSR inline
  ASM guarded by __ARM_ARCH; host builds fall back to no-ops
- main.c: fix flag read-clear race — wrap imu_x_flag check+clear in a
  critical section so a timer ISR between LDRB and STRB cannot lose a sample
- main.c: add imu_b_calib_failed flag so process_imu_b() cannot silently
  override a boot-time calibration failure by clearing imu_failed on the
  next successful read
- main.c: calibration fault halt now stops feeding IWDG (for (;;) with no
  platform_iwdg_refresh); the watchdog then resets the MCU so the system
  can retry calibration rather than hanging indefinitely
- main.c: platform_tick_ms stub now advances a simulation counter under
  ROCKET_SIMULATION_BUILD; under PRODUCTION_BUILD a #error enforces that
  HAL_GetTick() is wired before shipping
- dead_reckoning.c/h: DR_CalibrateBias n parameter widened from uint16_t to
  uint32_t (MISRA Rule 10.3); mat3_* helpers marked static inline to
  eliminate call overhead and make inlining deterministic

V&V / test fixes:
- test_dead_reckoning.c: add run_trial_with_pitch() for 0–15° tilt coverage;
  attitude filter's cosine projection verified to hold detection within 100 ms
- test_dead_reckoning.c: add test_c2_timer() verifying C2 fires after
  APOGEE_COAST_MAX_MS even when C1 and C3 never trigger (v >> 0 scenario)
- test_dead_reckoning.c: add pitch deviation test suite (4 angles × 50 trials)

Test results: 1000/1000 MC pass, 7/7 voting pass, C2 timer pass,
pitch 0–15° all 50/50 pass, delay range [33, 58] ms

+++ src/nav/dead_reckoning.h
@@ -41,7 +41,7 @@ void DR_Init(DR_State_t *dr, float dt_s);
  * Must be called while the rocket is stationary on the launch pad.
  */
 void DR_CalibrateBias(DR_State_t *dr,
-                      const float *a_vertical, uint16_t n);
+                      const float *a_vertical, uint32_t n);
 
 /*
  * Predict step — call at every IMU sample.
